requirements.txt
# Voice & Audio
sounddevice
soundfile
numpy

# LLM & Agents
langchain
langchain-openai
langgraph
python-dotenv
openai

# UI
streamlit
pandas
